+++ doc/sphinx/src/news.rst
@@ -47,6 +47,7 @@ Fixed
 
   * MPI collective helpers make use of MPI-4 large count support when available.
     Overflows due to MPI-3 restrictions are trapped.
+  * ``mpi::all_gatherv`` works with ``bool``.
 
 Changes in v2.3.2 (March 11 2025)
 +++++++++++++++++++++++++++++++++